fixup: rename contact models file
