+++ src/CabanaPD.hpp
@@ -29,7 +29,7 @@
 #include <CabanaPD_Types.hpp>
 #include <CabanaPD_config.hpp>
 
-#include <force/CabanaPD_ForceModels_Contact.hpp>
+#include <force/CabanaPD_ContactModels.hpp>
 #include <force/CabanaPD_ForceModels_LPS.hpp>
 #include <force/CabanaPD_ForceModels_PMB.hpp>
 #include <force/CabanaPD_Force_Contact.hpp>
